// express/schemas/schema.d.ts
export interface components {
  schemas: {
// ...
    CreateArtist: {
      pictureIdUpload: components['schemas']['Id']
      name: components['schemas']['Name']
      description: components['schemas']['Description']
    }
    UpdateArtist: components['schemas']['CreateArtist']
    ListArtists: {
      idArtist: components['schemas']['Id']
      pictureUrl: components['schemas']['Url']
      name: components['schemas']['Name']
    }[]
// ...
    GetEvent: {
// ...
      images: {
        idUpload: components['schemas']['Id']
        url: components['schemas']['Url']
      }[]
      artists: components['schemas']['ListArtists']
    }
// ...
    CreateReview: {
      title: components['schemas']['Name']
      eventLocation: components['schemas']['LongName']
      eventDate: components['schemas']['Date']
      content: components['schemas']['Content']
      rate: components['schemas']['Rate']
    }
// ...
  }
  responses: {
    /** @description Invalid request contents */
    BadRequest: {
      content: {
        'application/json': components['schemas']['Error']
      }
    }
// ...
  }
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}

export type $defs = Record<string, never>

export type external = Record<string, never>

export interface operations {
  /** @description List available artists */
  ListArtists: {
    parameters: {
      query?: {
        /** @description A string to search artists by name */
        search?: string
        /** @description A listing page number */
        page?: components['schemas']['Page']
      }
    }
    responses: {
      /** @description Available artists */
      200: {
        content: {
          'application/json': {
            count: number
            pages: number
            currentPage: number
            results: components['schemas']['ListArtists']
          }
        }
      }
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Create new artist */
  CreateArtist: {
    requestBody?: {
      content: {
        'application/json': components['schemas']['CreateArtist']
      }
    }
    responses: {
      /** @description Created artist details */
      201: {
        content: {
          'application/json': components['schemas']['GetArtist']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  UpdateArtist: {
    parameters: {
      path: {
        /** @description The unique identifier of the artist */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': components['schemas']['UpdateArtist']
      }
    }
    responses: {
      /** @description Artist updated details */
      200: {
        content: {
          'application/json': components['schemas']['GetArtist']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      403: components['responses']['Forbidden']
      default: components['responses']['InternalServerError']
    }
  }
  /** @description Create artist's performance review */
  CreateReview: {
    parameters: {
      path: {
        /** @description The unique identifier of the artist */
        id: components['schemas']['Id']
      }
    }
    requestBody?: {
      content: {
        'application/json': components['schemas']['CreateReview']
      }
    }
    responses: {
      /** @description Created review details */
      201: {
        content: {
          'application/json': components['schemas']['GetReview']
        }
      }
      400: components['responses']['BadRequest']
      401: components['responses']['Unauthorized']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
// ...
  SignUp: {
    requestBody?: {
      content: {
        'application/json': {
          signUpToken: components['schemas']['JWT']
          email: components['schemas']['Email']
          username: components['schemas']['Name']
          name: components['schemas']['Name']
          surname: components['schemas']['Name']
          birthdate: components['schemas']['Date']
        }
      }
    }
    responses: {
      /** @description Successfully signed up user account */
      204: {
        content: never
      }
      400: components['responses']['BadRequest']
      429: components['responses']['TooManyRequests']
      default: components['responses']['InternalServerError']
    }
  }
// ...
}